Allow configure dialog to open without editor.
Addresses #910.

The solution is a bit of a hack. The offending checkbox should probably
be completely removed or disabled if there is no editor (at the moment
it can be toggled, but won't have any effect). Also the if-statements
introduced in this commit are a bit of a hack. There is probably some
better way to address this without the repeated checking whether
Nengo.ace is defined. However, it might make sense to wait for the
completion of the frontend refactoring before addressing this.